ImageSet.geLabels(): Compute seed matrix for labelTemplate
The 'labelTemplate' function now accepts a seed matrix, which has '0'
in background pixels, '1' in the object region (area containing the
object, but also contains other pixels), and '2' in the keypoints which
are known to be on the object in the test image.

get getLabels() function now creates this seed matrix and passes it to
labelTemplates() which is currently still a stub.