labelTemplate improvements
Improved the comparison between the warped template and the image.
The image region's indices are now calculated correctly to prevent
out-of-bounds bugs.
The warped template is now cropped to make sure it's the same size as
the respective image region.
Comparison is done with gradient magnitude and direction.

Morphological operations are now applied to the final label image in
order to remove small holes from it.